Move encoding methods to calls_encoding.py (#418)
Moves 6 methods to the new CallsEncodingMixin:
- Base64: base64_encode, base64_decode
- URL: url_encode, url_decode, url_parse, url_join

All emit heap-allocating state machines in WAT and return
Result/UrlParts ADTs.

calls.py: 5,168 -> 3,135 lines (-2,033)
calls_encoding.py: 14 -> 2,048 lines
